requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
joblib==1.0.1
jsonpickle==2.0.0
nltk==3.6.2
regex==2021.4.4
requests==2.25.1
smmap==4.0.0
tqdm==4.61.1
urllib3==1.26.5
